+++ auth/detect/internal/externalaccount/awsprovider.go
