+++ auth/detect/internal/externalaccount/awsprovider_test.go
