+++ auth/detect/internal/externalaccount/executableprovider.go
