+++ auth/detect/internal/externalaccount/fileprovider_test.go
